@@ -1,20 +1,25 @@
 name: zeta_flutter
-description: Zeta
+description: Zebra Design System (Zeta) - Flutter Component Library
 version: 0.0.1+10
-homepage:
+repository: "https:/github.com/zebradevs/zeta-flutter"
+
+platforms:
+  android:
+  ios:
+  macos:
+  web:
+  windows:
+  linux:
 
 environment:
-  sdk: ">=2.19.1 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=3.0.1 <4.0.0"
+  flutter: ">=3.7.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  dart_code_metrics: ^5.6.0
-  dart_code_metrics_presets: ^1.4.1
-  flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
 
