@@ -1,8 +1,7 @@
-// ignore_for_file: public_member_api_docs
-
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:zeta_flutter/zeta_flutter.dart';
+import 'pages/color_example.dart';
 import 'pages/grid_example.dart';
 import 'pages/spacing_example.dart';
 import 'pages/typography_example.dart';
@@ -29,6 +28,7 @@ final List<Component> components = [
   Component(GridExample.name, const GridExample()),
   Component(SpacingExample.name, const SpacingExample()),
   Component(TypographyExample.name, const TypographyExample()),
+  Component(ColorExample.name, const ColorExample()),
 ];
 
 class _MyAppState extends State<MyApp> {
@@ -38,15 +38,13 @@ class _MyAppState extends State<MyApp> {
         path: '/',
         builder: (_, __) => const Home(),
         routes: [
-          ...components
-              .map(
-                (e) => GoRoute(
-                  path: e.name,
-                  builder: (_, __) => e.page,
-                  routes: e.children.map((f) => GoRoute(path: f.name, builder: (_, __) => f.page)).toList(),
-                ),
-              )
-              .toList()
+          ...components.map(
+            (e) => GoRoute(
+              path: e.name,
+              builder: (_, __) => e.page,
+              routes: e.children.map((f) => GoRoute(path: f.name, builder: (_, __) => f.page)).toList(),
+            ),
+          )
         ],
       ),
     ],
@@ -54,30 +52,43 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp.router(
-      theme: ZetaTheme.zeta,
-      routerConfig: router,
+    return Zeta(
+      builder: (context, theme, colors) {
+        return MaterialApp.router(theme: theme, routerConfig: router);
+      },
     );
   }
 }
 
-class Home extends StatelessWidget {
+class Home extends StatefulWidget {
   const Home({super.key});
 
   @override
-  Widget build(BuildContext context) {
-    final items = components..sort(((a, b) => a.name.compareTo(b.name)));
+  State<Home> createState() => _HomeState();
+}
 
+class _HomeState extends State<Home> {
+  @override
+  Widget build(BuildContext context) {
+    final items = components..sort((a, b) => a.name.compareTo(b.name));
+
+    final colors = ZetaColors.of(context);
     return Scaffold(
-      appBar: AppBar(title: const Text('Zeta')),
-      body: ListView.builder(
-        itemCount: items.length,
-        itemBuilder: (context, index) {
-          return ListTile(
-            title: Text(items[index].name),
-            onTap: () => context.go('/${items[index].name}'),
-          );
-        },
+      appBar: AppBar(
+        title: const Text('Zeta'),
+        backgroundColor: colors.primary,
+      ),
+      body: ColoredBox(
+        color: colors.background,
+        child: ListView.builder(
+          itemCount: items.length,
+          itemBuilder: (context, index) {
+            return ListTile(
+              title: ZetaText(items[index].name),
+              onTap: () => context.go('/${items[index].name}'),
+            );
+          },
+        ),
       ),
     );
   }
