@@ -25,36 +25,44 @@ WidgetbookComponent textWidgetBook() {
       ),
       WidgetbookUseCase(
         name: 'Universal sizes',
-        builder: (context) => ZetaText(
-          exampleText,
-          fontSize: context.knobs.options(
-            label: 'Sizes',
-            labelBuilder: (p0) => p0 == 14 ? 'x3_5' : 'x${p0! ~/ 4}',
-            options: const [
-              Dimensions.x3,
-              Dimensions.x3_5,
-              Dimensions.x4,
-              Dimensions.x5,
-              Dimensions.x6,
-              Dimensions.x7,
-              Dimensions.x8,
-              Dimensions.x9,
-              Dimensions.x10,
-              Dimensions.x11,
-              Dimensions.x12,
-              Dimensions.x13,
-            ],
+        builder: (context) => Container(
+          color: ZetaColors.of(context).background,
+          padding: const EdgeInsets.all(Dimensions.l),
+          child: ZetaText(
+            exampleText,
+            fontSize: context.knobs.list(
+              label: 'Sizes',
+              labelBuilder: (p0) => p0 == 14 ? 'x3_5' : 'x${p0! ~/ 4}',
+              options: const [
+                Dimensions.x3,
+                Dimensions.x3_5,
+                Dimensions.x4,
+                Dimensions.x5,
+                Dimensions.x6,
+                Dimensions.x7,
+                Dimensions.x8,
+                Dimensions.x9,
+                Dimensions.x10,
+                Dimensions.x11,
+                Dimensions.x12,
+                Dimensions.x13,
+              ],
+            ),
           ),
         ),
       ),
       WidgetbookUseCase(
         name: 'Dedicated sizes',
-        builder: (context) => ZetaText(
-          exampleText,
-          style: context.knobs.options(
-            label: 'Sizes',
-            labelBuilder: (p0) => dedicatedSizes.entries.firstWhere((element) => element.value == p0).key,
-            options: dedicatedSizes.values.toList(),
+        builder: (context) => Container(
+          color: ZetaColors.of(context).background,
+          padding: const EdgeInsets.all(Dimensions.l),
+          child: ZetaText(
+            exampleText,
+            style: context.knobs.list(
+              label: 'Sizes',
+              labelBuilder: (p0) => dedicatedSizes.entries.firstWhere((element) => element.value == p0).key,
+              options: dedicatedSizes.values.toList(),
+            ),
           ),
         ),
       ),
@@ -67,44 +75,42 @@ class _TextWrapper extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ZetaText(
-      context.knobs.text(label: 'Input text', initialValue: exampleText.split(',').first),
-      decoration: context.knobs.boolean(label: 'Underline') ? TextDecoration.underline : null,
-      first: context.knobs.boolean(label: 'First'),
-      last: context.knobs.boolean(label: 'Last'),
-      fontStyle: context.knobs.boolean(label: 'Italic') ? FontStyle.italic : null,
-      fontWeight: context.knobs.options(
-        label: 'Font Weight',
-        labelBuilder: (p0) => p0 == FontWeight.w400
-            ? 'Default'
-            : p0 == FontWeight.w300
-                ? 'Light'
-                : 'Medium',
-        options: const [
-          FontWeight.w400,
-          FontWeight.w300,
-          FontWeight.w500,
-        ],
+    final ZetaColors colors = ZetaColors.of(context);
+    return Container(
+      color: colors.background,
+      padding: const EdgeInsets.all(Dimensions.l),
+      child: ZetaText(
+        context.knobs.string(label: 'Input text', initialValue: exampleText.split(',').first),
+        decoration: context.knobs.boolean(label: 'Underline') ? TextDecoration.underline : null,
+        first: context.knobs.boolean(label: 'First'),
+        last: context.knobs.boolean(label: 'Last'),
+        fontStyle: context.knobs.boolean(label: 'Italic') ? FontStyle.italic : null,
+        fontWeight: context.knobs.list(
+          label: 'Font Weight',
+          labelBuilder: (p0) => p0 == FontWeight.w400
+              ? 'Default'
+              : p0 == FontWeight.w300
+                  ? 'Light'
+                  : 'Medium',
+          options: const [FontWeight.w400, FontWeight.w300, FontWeight.w500],
+        ),
+        resetHeight: context.knobs.boolean(label: 'Reset height'),
+        textColor: context.knobs.list(
+          label: 'Text color',
+          labelBuilder: (p0) => p0?.value == colors.textDefault.value ? 'Default' : 'Subtle',
+          options: [colors.textDefault, colors.textSubtle],
+        ),
+        textDirection: context.knobs.list(
+          label: 'Text direction',
+          options: const [
+            TextDirection.ltr,
+            TextDirection.rtl,
+          ],
+        ),
+        upperCase: context.knobs.boolean(label: 'Upper case'),
+        maxWidth: context.knobs.double.slider(label: 'Width', initialValue: 66, max: 100, min: 10, divisions: 90),
+        fontSize: context.knobs.double.slider(label: 'Font size', initialValue: 12, divisions: 42, min: 12, max: 96),
       ),
-      resetHeight: context.knobs.boolean(label: 'Reset height'),
-      textColor: context.knobs.options(
-        label: 'Text color',
-        labelBuilder: (p0) => p0 == ZetaColors.textColor ? 'Default' : 'Subtle',
-        options: const [
-          ZetaColors.textColor,
-          ZetaColors.textColorSubtle,
-        ],
-      ),
-      textDirection: context.knobs.options(
-        label: 'Text direction',
-        options: const [
-          TextDirection.ltr,
-          TextDirection.rtl,
-        ],
-      ),
-      upperCase: context.knobs.boolean(label: 'Upper case'),
-      maxWidth: context.knobs.slider(label: 'Width', initialValue: 66, max: 100, min: 10, divisions: 90),
-      fontSize: context.knobs.slider(label: 'Font size', initialValue: 12, divisions: 42, min: 12, max: 96),
     );
   }
 }
